fixed Windows build issues; OUI database added to Windows installer

+++ include/nms_agent.h
@@ -1262,8 +1262,8 @@ struct LIBNXAGENT_EXPORTABLE ProcessInformation
 /**
  * API for subagents
  */
-bool LIBNXAGENT_EXPORTABLE AgentGetParameterArgA(const TCHAR *param, int index, char *arg, int maxSize, bool inBrackets = true);
-bool LIBNXAGENT_EXPORTABLE AgentGetParameterArgW(const TCHAR *param, int index, WCHAR *arg, int maxSize, bool inBrackets = true);
+bool LIBNXAGENT_EXPORTABLE AgentGetParameterArgA(const TCHAR *param, int index, char *arg, size_t maxSize, bool inBrackets = true);
+bool LIBNXAGENT_EXPORTABLE AgentGetParameterArgW(const TCHAR *param, int index, WCHAR *arg, size_t maxSize, bool inBrackets = true);
 #ifdef UNICODE
 #define AgentGetParameterArg AgentGetParameterArgW
 #else
